+++ spec/lib/ember_cli/runner_spec.rb
@@ -3,20 +3,37 @@ require "ember_cli/runner"
 describe EmberCli::Runner do
   describe "#run!" do
     context "when the command fails" do
-      it "writes to all `err` and `out` streams" do
-        output_streams = Array.new(4) { StringIO.new }
+      it "writes output to `out` streams" do
+        output_streams = Array.new(2) { StringIO.new }
         runner = EmberCli::Runner.new(
-          err: output_streams.first(2),
-          out: output_streams.last(2),
+          err: [],
+          out: output_streams,
         )
 
         expect { runner.run!("echo 'out'; echo 'err' > /dev/stderr; exit 1") }.
           to raise_error(SystemExit)
 
-        output_streams.each(&:rewind)
+        ouput_strings = output_streams.each(&:rewind).map(&:read)
 
-        output_streams.each do |stream|
-          expect(stream.read).to match(/out\nerr/)
+        ouput_strings.each do |output|
+          expect(output).to eq("out\n")
+        end
+      end
+
+      it "writes errors to `err` streams" do
+        error_streams = Array.new(2) { StringIO.new }
+        runner = EmberCli::Runner.new(
+          err: error_streams,
+          out: [],
+        )
+
+        expect { runner.run!("echo 'out'; echo 'err' > /dev/stderr; exit 1") }.
+          to raise_error(SystemExit)
+
+        error_strings = error_streams.each(&:rewind).map(&:read)
+
+        error_strings.each do |error|
+          expect(error).to eq("err\n")
         end
       end
     end
@@ -26,10 +43,11 @@ describe EmberCli::Runner do
       err = StringIO.new
       runner = EmberCli::Runner.new(err: [err], out: [out])
 
-      runner.run!("echo 'out'")
+      status = runner.run!("echo 'out'")
 
       [err, out].each(&:rewind)
 
+      expect(status).to be_success
       expect(err.read).to be_empty
       expect(out.read).to eq("out\n")
     end